+++ Hackathon.Common/Models/Friend/FriendshipStatus.cs
@@ -1,17 +1,17 @@
 namespace Hackathon.Common.Models.Friend;
 
-public enum FriendshipStatus
+public enum FriendshipStatus: byte
 {
     /// <summary>
     /// Предлоджение дружбы отправлено
     /// </summary>
     Pending = 1,
-    
+
     /// <summary>
     /// Предложение дружбы принято
     /// </summary>
     Confirmed = 2,
-    
+
     /// <summary>
     /// Предложение дружбы отклонено
     /// </summary>
