+++ Hackathon.UI/src/app/app-routing.module.ts
@@ -5,10 +5,8 @@ import {RegisterComponent} from "./components/register/register.component";
 import {NotFoundComponent} from "./components/not-found/not-found.component";
 import {AuthGuard} from "./services/auth.guard";
 import {EventListComponent} from "./components/event/list/event.list.component";
-import {EventViewComponent} from "./components/event/view/event.view.component";
 import {EventFormComponent} from "./components/event/form/event.form.component";
 import {UserListComponent} from "./components/user/list/user.list.component";
-import {UserViewComponent} from "./components/user/view/user.view.component";
 import {TeamListComponent} from "./components/team/list/team.list.component";
 import {TeamNewComponent} from "./components/team/new/team.new.component";
 import {NotificationListComponent} from "./components/notification/list/notification.list.component";
@@ -37,7 +35,7 @@ const routes: Routes = [
   { path: 'team/:teamId', component: TeamViewComponent, canActivate: [AuthGuard] },
 
   { path: 'users', component: UserListComponent, canActivate: [AuthGuard] },
-  { path: 'users/:userId', component: UserViewComponent, canActivate: [AuthGuard] },
+  { path: 'users/:userId', component: ProfileViewComponent, canActivate: [AuthGuard] },
   { path: 'notifications', component: NotificationListComponent, canActivate: [AuthGuard]},
 
   { path: 'eventLog', component: EventLogComponent, canActivate: [AuthGuard] },
