+++ Hackathon.UI/src/app/components/profile/image/profile-image.component.scss
@@ -1,6 +1,5 @@
 :host {
 
-  cursor: pointer;
   position: relative;
   display: flex;
   flex-direction: column;
@@ -55,4 +54,14 @@
   &:hover button {
     opacity: 1;
   }
+}
+
+.noselect {
+  -webkit-touch-callout: none; /* iOS Safari */
+  -webkit-user-select: none; /* Safari */
+  -khtml-user-select: none; /* Konqueror HTML */
+  -moz-user-select: none; /* Old versions of Firefox */
+  -ms-user-select: none; /* Internet Explorer/Edge */
+  user-select: none; /* Non-prefixed version, currently
+                                  supported by Chrome, Edge, Opera and Firefox */
 }
