+++ Hackathon.UI/src/app/components/profile/view/profile.view.component.scss
@@ -7,5 +7,36 @@
   }
 }
 
+.reaction__list
+{
+  display: flex;
+  padding-top: 1em;
+  border-top: 1px solid #f5f5f5;
+
+  .reaction
+  {
+    display: block;
+    height: 40px;
+    width: 40px;
+    cursor: pointer;
+
+    color: #b8b8b8;
+
+    &.reaction-like {
+      &:hover { color: #2886c9; }
+      &.active { color: #2886c9; }
+    }
+
+    &.reaction-heart {
+      &:hover { color: #da2121; }
+      &.active { color: #da2121; }
+    }
+
+    &.reaction-fire {
+      &:hover { color: #f1522c; }
+      &.active { color: #f1522c; }
+    }
+  }
+}
