+++ Hackathon.UI/src/app/components/toolbar/toolbar.component.html
@@ -5,7 +5,7 @@
   <span class="spacer"></span>
   <notification-bell></notification-bell>
   <button mat-icon-button [matMenuTriggerFor]="profileMenu">
-    <profile-image [style.width]="'40px'" [style.height]="'40px'" [style.font-size]="'18px'"></profile-image>
+    <profile-image [userId]="userId ?? 0" [style.width]="'40px'" [style.height]="'40px'" [style.font-size]="'18px'"></profile-image>
   </button>
   <mat-menu #profileMenu="matMenu" yPosition="below">
     <button mat-menu-item routerLink="/profile">
