--- Hackathon.UI/src/app/components/user/view/user.view.component.html
@@ -1,36 +1,0 @@
-<layout-default title="Пользователь" [isLoading]="isLoading">
-
-  <mat-tab-group *ngIf="user" class="vertical-tabs">
-    <mat-tab label="Общая информация">
-
-      <list-details [items]="[
-        { key: 'Имя пользователя', value: user?.userName},
-        { key: 'Полное имя', value: user?.fullName},
-        { key: 'Email', value: user?.email},
-      ]"></list-details>
-
-      <div class="reaction__list">
-
-        <div class="reaction reaction-{{ UserProfileReaction[reaction].toLowerCase() }}"
-             (click)="toggleReaction($event, reaction)"
-             [class.active]="isReactionActive(reaction)"
-             *ngFor="let reaction of getAvailableUserProfileReactions()">
-
-          <div [ngSwitch]="reaction">
-            <mat-icon class="reaction-icon" *ngSwitchCase="UserProfileReaction.Like">thumb_up</mat-icon>
-            <mat-icon class="reaction-icon" *ngSwitchCase="UserProfileReaction.Heart">favorite</mat-icon>
-            <mat-icon class="reaction-icon" *ngSwitchCase="UserProfileReaction.Fire">local_fire_department</mat-icon>
-          </div>
-
-      </div>
-
-      </div>
-
-    </mat-tab>
-  </mat-tab-group>
-
-</layout-default>
-
-<ng-template>
-
-</ng-template>
